Robust Backfitting

This function computes a robust backfitting algorithm for additive models

backf.rob(
  formula,
  data,
  subset,
  windows,
  point = NULL,
  epsilon = 1e-06,
  degree = 0,
  sigma.hat = NULL,
  prob = NULL,
  max.it = 50,
  k.h = 1.345,
  k.t = 4.685,
  type = "Huber"
)

Arguments

formula
an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which the function was called.

subset
an optional vector specifying a subset of observations to be used in the fitting process.

windows
vector of bandwidths for the local polynomial smoother, one per explanatory variable.

point
matrix of points where predictions will be computed and returned.

epsilon
convergence criterion. Maximum allowed relative difference between consecutive estimates

degree
degree of the local polynomial smoother. Defaults to 0 (local constant).

sigma.hat
estimate of the residual standard error. If NULL (default) we use the mad of the residuals obtained with local medians.

prob
vector of probabilities of observing each response (length n). Defaults to NULL and in that case it is ignored.

max.it
Maximum number of iterations for the algorithm.

k.h
tuning constant for a Huber-type loss function.

k.t
tuning constant for a Tukey-type loss function.

type
one of either 'Tukey' or 'Huber'.

Value

A list with the following components:

alpha

Estimate for the intercept.

g.matrix

Matrix of estimated additive components (n by p).

prediction

Matrix of estimated additive components for the points listed in the argument point.

sigma.hat

Estimate of the residual standard error.

Details

This function computes a robust backfitting algorithm for additive models using robust local polynomial smoothers.

References

Boente G, Martinez A, Salibian-Barrera M. Robust estimators for additive models using backfitting. Journal of Nonparametric Statistics, 2017; 29:744-767. https://doi.org/10.1080/10485252.2017.1369077

Author

Matias Salibian-Barrera, matias@stat.ubc.ca, Alejandra Martinez

Examples

data(airquality)
tmp <- backf.rob(Ozone ~ Solar.R + Wind + Temp, data=airquality, 
subset=complete.cases(airquality), windows=c(136.7, 8.9, 4.8), degree=1)